Final Score now resets properly on game restart.
Missile time alive and slightly changed speed so it can actually catch the Booster Bot when it is going side-to-side.

Changed Enemy movement patters around, now make a lot more sense, easiest has the easiest movement, hardest has the hardest :)

Added model tilting on movement.

Ability to rotate the players with the right joystick. There is a little problem that occurs when separating the players rotations, where when you are facing up, you are your fastest and when facing down you are your slowest. Fuck if I know what's up.

PlayerManagement script is getting massive, so I made the effort to comment heaps of stuff and clean up unneeded variables and sections that had been commented out for ages that were not really needed anymore and were just references.